refactor(plugin/evm): move atomic sync code back to plugin/evm
