+++ plugin/evm/atomic/sync/leaf_handler.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2019-2025, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package atomic
+package sync
 
 import (
 	"context"
