+++ plugin/evm/atomic/sync/summary_provider.go
@@ -1,6 +1,6 @@
 // Copyright (C) 2019-2025, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
-package atomic
+package sync
 
 import (
 	"fmt"
